Separate module for trait object functions from global static functions
