+++ crates/common/src/msgbus/mod.rs
@@ -19,6 +19,7 @@
 //! - Pub/Sub
 //! - Request/Response
 
+pub mod core;
 pub mod database;
 pub mod handler;
 pub mod listener;
@@ -30,24 +31,13 @@ pub mod switchboard;
 #[cfg(test)]
 mod tests;
 
-use std::{
-    any::Any,
-    cell::RefCell,
-    collections::{BTreeSet, HashMap},
-    fmt::{self, Debug, Display},
-    hash::{Hash, Hasher},
-    ops::Deref,
-    rc::Rc,
-    sync::OnceLock,
-};
+pub use core::MessageBus;
+use std::{self, any::Any, cell::RefCell, fmt::Debug, rc::Rc, sync::OnceLock};
 
-use ahash::AHashMap;
 use handler::ShareableMessageHandler;
-use indexmap::IndexMap;
 use matching::is_matching_backtracking;
 use nautilus_core::UUID4;
-use nautilus_model::{data::Data, identifiers::TraderId};
-use switchboard::MessagingSwitchboard;
+use nautilus_model::data::Data;
 use ustr::Ustr;
 
 use crate::messages::data::DataResponse;
@@ -203,14 +193,14 @@ pub fn subscribe<T: AsRef<str>>(
     handler: ShareableMessageHandler,
     priority: Option<u8>,
 ) {
-    let pattern = Pattern::from(pattern);
+    let pattern = core::Pattern::from(pattern);
     inner_subscribe(pattern, handler, priority);
 }
 
-fn inner_subscribe(pattern: Pattern, handler: ShareableMessageHandler, priority: Option<u8>) {
+fn inner_subscribe(pattern: core::Pattern, handler: ShareableMessageHandler, priority: Option<u8>) {
     let msgbus = get_message_bus();
     let mut msgbus_ref_mut = msgbus.borrow_mut();
-    let sub = Subscription::new(pattern, handler, priority);
+    let sub = core::Subscription::new(pattern, handler, priority);
 
     log::debug!(
         "Subscribing {:?} for pattern '{}'",
@@ -238,14 +228,14 @@ fn inner_subscribe(pattern: Pattern, handler: ShareableMessageHandler, priority:
 
 /// Unsubscribes the `handler` from the `pattern`.
 pub fn unsubscribe<T: AsRef<str>>(pattern: T, handler: ShareableMessageHandler) {
-    let pattern = Pattern::from(pattern);
+    let pattern = core::Pattern::from(pattern);
     inner_unsubscribe(pattern, handler);
 }
 
-fn inner_unsubscribe(pattern: Pattern, handler: ShareableMessageHandler) {
+fn inner_unsubscribe(pattern: core::Pattern, handler: ShareableMessageHandler) {
     log::debug!("Unsubscribing {handler:?} from pattern '{pattern}'");
 
-    let sub = Subscription::new(pattern, handler, None);
+    let sub = core::Subscription::new(pattern, handler, None);
 
     get_message_bus()
         .borrow_mut()
@@ -267,389 +257,11 @@ fn inner_unsubscribe(pattern: Pattern, handler: ShareableMessageHandler) {
 }
 
 pub fn is_subscribed<T: AsRef<str>>(pattern: T, handler: ShareableMessageHandler) -> bool {
-    let pattern = Pattern::from(pattern);
-    let sub = Subscription::new(pattern, handler, None);
+    let pattern = core::Pattern::from(pattern);
+    let sub = core::Subscription::new(pattern, handler, None);
     get_message_bus().borrow().subscriptions.contains(&sub)
 }
 
 pub fn subscriptions_count<T: AsRef<str>>(topic: T) -> usize {
     get_message_bus().borrow().subscriptions_count(topic)
-}
-
-/// A string pattern for a subscription. The pattern is used to match topics.
-///
-/// A pattern is made of characters:
-/// - `*` - match 0 or more characters
-/// - `?` - match any character once
-/// - `a-z` - match the specific character
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
-pub struct Pattern(pub Ustr);
-
-impl<T: AsRef<str>> From<T> for Pattern {
-    fn from(value: T) -> Self {
-        Self(Ustr::from(value.as_ref()))
-    }
-}
-
-impl Deref for Pattern {
-    type Target = Ustr;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl Display for Pattern {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl From<Topic> for Pattern {
-    fn from(value: Topic) -> Self {
-        Self(value.0)
-    }
-}
-
-/// A string topic for publishing data. It is a fully qualified pattern i.e.
-/// wildcard characters are not allowed.
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
-pub struct Topic(pub Ustr);
-
-impl Topic {
-    pub fn validate<T: AsRef<str>>(topic: T) -> bool {
-        let topic = Ustr::from(topic.as_ref());
-        !topic.chars().any(|c| c == '*' || c == '?')
-    }
-}
-
-impl<T: AsRef<str>> From<T> for Topic {
-    fn from(value: T) -> Self {
-        Self(Ustr::from(value.as_ref()))
-    }
-}
-
-impl Deref for Topic {
-    type Target = Ustr;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl Display for Topic {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-/// Represents a subscription to a particular topic.
-///
-/// This is an internal class intended to be used by the message bus to organize
-/// topics and their subscribers.
-///
-
-#[derive(Clone, Debug)]
-pub struct Subscription {
-    /// The shareable message handler for the subscription.
-    pub handler: ShareableMessageHandler,
-    /// Store a copy of the handler ID for faster equality checks.
-    pub handler_id: Ustr,
-    /// The pattern for the subscription.
-    pub pattern: Pattern,
-    /// The priority for the subscription determines the ordering of handlers receiving
-    /// messages being processed, higher priority handlers will receive messages before
-    /// lower priority handlers.
-    pub priority: u8,
-}
-
-impl Subscription {
-    /// Creates a new [`Subscription`] instance.
-    #[must_use]
-    pub fn new(pattern: Pattern, handler: ShareableMessageHandler, priority: Option<u8>) -> Self {
-        Self {
-            handler_id: handler.0.id(),
-            pattern,
-            handler,
-            priority: priority.unwrap_or(0),
-        }
-    }
-}
-
-impl PartialEq<Self> for Subscription {
-    fn eq(&self, other: &Self) -> bool {
-        self.pattern == other.pattern && self.handler_id == other.handler_id
-    }
-}
-
-impl Eq for Subscription {}
-
-impl PartialOrd for Subscription {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for Subscription {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        other.priority.cmp(&self.priority)
-    }
-}
-
-impl Hash for Subscription {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.pattern.hash(state);
-        self.handler_id.hash(state);
-    }
-}
-
-/// A generic message bus to facilitate various messaging patterns.
-///
-/// The bus provides both a producer and consumer API for Pub/Sub, Req/Rep, as
-/// well as direct point-to-point messaging to registered endpoints.
-///
-/// Pub/Sub wildcard patterns for hierarchical topics are possible:
-///  - `*` asterisk represents one or more characters in a pattern.
-///  - `?` question mark represents a single character in a pattern.
-///
-/// Given a topic and pattern potentially containing wildcard characters, i.e.
-/// `*` and `?`, where `?` can match any single character in the topic, and `*`
-/// can match any number of characters including zero characters.
-///
-/// The asterisk in a wildcard matches any character zero or more times. For
-/// example, `comp*` matches anything beginning with `comp` which means `comp`,
-/// `complete`, and `computer` are all matched.
-///
-/// A question mark matches a single character once. For example, `c?mp` matches
-/// `camp` and `comp`. The question mark can also be used more than once.
-/// For example, `c??p` would match both of the above examples and `coop`.
-#[cfg_attr(
-    feature = "python",
-    pyo3::pyclass(module = "nautilus_trader.core.nautilus_pyo3.common")
-)]
-#[derive(Debug)]
-pub struct MessageBus {
-    /// The trader ID associated with the message bus.
-    pub trader_id: TraderId,
-    /// The instance ID associated with the message bus.
-    pub instance_id: UUID4,
-    /// The name for the message bus.
-    pub name: String,
-    /// If the message bus is backed by a database.
-    pub has_backing: bool,
-    /// The switchboard for built-in endpoints.
-    pub switchboard: MessagingSwitchboard,
-    /// Active subscriptions.
-    subscriptions: BTreeSet<Subscription>,
-    /// Maps a topic to all the handlers registered for it
-    /// this is updated whenever a new subscription is created.
-    topics: IndexMap<Topic, Vec<Subscription>>,
-    /// Index of endpoint addresses and their handlers.
-    endpoints: IndexMap<Ustr, ShareableMessageHandler>,
-    /// Index of request correlation IDs and their response handlers.
-    correlation_index: AHashMap<UUID4, ShareableMessageHandler>,
-}
-
-// SAFETY: Message bus is not meant to be passed between threads
-#[allow(unsafe_code)]
-unsafe impl Send for MessageBus {}
-#[allow(unsafe_code)]
-unsafe impl Sync for MessageBus {}
-
-impl MessageBus {
-    /// Creates a new [`MessageBus`] instance.
-    #[must_use]
-    pub fn new(
-        trader_id: TraderId,
-        instance_id: UUID4,
-        name: Option<String>,
-        _config: Option<HashMap<String, serde_json::Value>>,
-    ) -> Self {
-        Self {
-            trader_id,
-            instance_id,
-            name: name.unwrap_or(stringify!(MessageBus).to_owned()),
-            switchboard: MessagingSwitchboard::default(),
-            subscriptions: BTreeSet::new(),
-            topics: IndexMap::new(),
-            endpoints: IndexMap::new(),
-            correlation_index: AHashMap::new(),
-            has_backing: false,
-        }
-    }
-
-    /// Returns the message bus instances memory address.
-    #[must_use]
-    pub fn memory_address(&self) -> String {
-        format!("{:?}", std::ptr::from_ref(self))
-    }
-
-    /// Returns the registered endpoint addresses.
-    #[must_use]
-    pub fn endpoints(&self) -> Vec<&str> {
-        self.endpoints.keys().map(Ustr::as_str).collect()
-    }
-
-    /// Returns actively subscribed patterns.
-    #[must_use]
-    pub fn patterns(&self) -> Vec<&str> {
-        self.subscriptions
-            .iter()
-            .map(|s| s.pattern.as_str())
-            .collect()
-    }
-
-    /// Returns whether there are subscribers for the given `topic`.
-    pub fn has_subscribers<T: AsRef<str>>(&self, topic: T) -> bool {
-        self.subscriptions_count(topic) > 0
-    }
-
-    /// Returns the count of subscribers for the given `topic`.
-    #[must_use]
-    pub fn subscriptions_count<T: AsRef<str>>(&self, topic: T) -> usize {
-        let topic = Topic::from(topic);
-        self.topics
-            .get(&topic)
-            .map(|subs| subs.len())
-            .unwrap_or_else(|| self.find_topic_matches(topic).len())
-    }
-
-    /// Returns active subscriptions.
-    #[must_use]
-    pub fn subscriptions(&self) -> Vec<&Subscription> {
-        self.subscriptions.iter().collect()
-    }
-
-    /// Returns the handler IDs for actively subscribed patterns.
-    #[must_use]
-    pub fn subscription_handler_ids(&self) -> Vec<&str> {
-        self.subscriptions
-            .iter()
-            .map(|s| s.handler_id.as_str())
-            .collect()
-    }
-
-    /// Returns whether there is a registered endpoint for the given `pattern`.
-    #[must_use]
-    pub fn is_registered<T: AsRef<str>>(&self, endpoint: T) -> bool {
-        self.endpoints.contains_key(&Ustr::from(endpoint.as_ref()))
-    }
-
-    /// Returns whether the given `handler` is subscribed to the given `pattern`.
-    #[must_use]
-    pub fn is_subscribed<T: AsRef<str>>(
-        &self,
-        pattern: T,
-        handler: ShareableMessageHandler,
-    ) -> bool {
-        let pattern = Pattern::from(pattern);
-        let sub = Subscription::new(pattern, handler, None);
-        self.subscriptions.contains(&sub)
-    }
-
-    /// Close the message bus which will close the sender channel and join the thread.
-    ///
-    /// # Errors
-    ///
-    /// This function never returns an error (TBD).
-    pub const fn close(&self) -> anyhow::Result<()> {
-        // TODO: Integrate the backing database
-        Ok(())
-    }
-
-    /// Returns the handler for the given `endpoint`.
-    #[must_use]
-    pub fn get_endpoint<T: AsRef<str>>(&self, endpoint: T) -> Option<&ShareableMessageHandler> {
-        self.endpoints.get(&Ustr::from(endpoint.as_ref()))
-    }
-
-    /// Returns the handler for the given `correlation_id`.
-    #[must_use]
-    pub fn get_response_handler(&self, correlation_id: &UUID4) -> Option<&ShareableMessageHandler> {
-        self.correlation_index.get(correlation_id)
-    }
-
-    /// Finds the subscriptions with pattern matching the given `topic`.
-    fn find_topic_matches(&self, topic: Topic) -> Vec<Subscription> {
-        self.subscriptions
-            .iter()
-            .filter_map(|sub| {
-                if is_matching_backtracking(topic, sub.pattern) {
-                    Some(sub.clone())
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    #[must_use]
-    /// Finds the subscriptions which match the given `topic` and caches the
-    /// results in the `patterns` map.
-    pub fn matching_subscriptions<T: AsRef<str>>(&mut self, topic: T) -> Vec<Subscription> {
-        let topic = Topic::from(topic);
-        self.inner_matching_subscriptions(topic)
-    }
-
-    fn inner_matching_subscriptions(&mut self, topic: Topic) -> Vec<Subscription> {
-        self.topics.get(&topic).cloned().unwrap_or_else(|| {
-            let mut matches = self.find_topic_matches(topic);
-            matches.sort();
-            self.topics.insert(topic, matches.clone());
-            matches
-        })
-    }
-
-    /// Register a response handler for a specific correlation ID.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if a handler is already registered for the given correlation ID.
-    pub fn register_response_handler(
-        &mut self,
-        correlation_id: &UUID4,
-        handler: ShareableMessageHandler,
-    ) -> anyhow::Result<()> {
-        if self.correlation_index.contains_key(correlation_id) {
-            return Err(anyhow::anyhow!(
-                "Correlation ID <{correlation_id}> already has a registered handler",
-            ));
-        }
-
-        self.correlation_index.insert(*correlation_id, handler);
-
-        Ok(())
-    }
-}
-
-/// Data specific functions.
-impl MessageBus {
-    // /// Send a [`DataRequest`] to an endpoint that must be a data client implementation.
-    // pub fn send_data_request(&self, message: DataRequest) {
-    //     // TODO: log error
-    //     if let Some(client) = self.get_client(&message.client_id, message.venue) {
-    //         let _ = client.request(message);
-    //     }
-    // }
-    //
-    // /// Send a [`SubscriptionCommand`] to an endpoint that must be a data client implementation.
-    // pub fn send_subscription_command(&self, message: SubscriptionCommand) {
-    //     if let Some(client) = self.get_client(&message.client_id, message.venue) {
-    //         client.through_execute(message);
-    //     }
-    // }
-
-    /// Register message bus globally
-    pub fn register_message_bus(self) -> Rc<RefCell<MessageBus>> {
-        let msgbus = Rc::new(RefCell::new(self));
-        set_message_bus(msgbus.clone());
-        msgbus
-    }
-}
-
-impl Default for MessageBus {
-    /// Creates a new default [`MessageBus`] instance.
-    fn default() -> Self {
-        Self::new(TraderId::from("TRADER-001"), UUID4::new(), None, None)
-    }
 }
